+++ CHANGELOG.md
@@ -6,11 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.6 - Unreleased]
 
 ### Added
 * Added progress bar in batch processing page.
 * Added option to erode particles by selected number of pixels in batch and single image processing pages.
+* Added option to fill holes inside particles in batch and single image processing pages.
+* Added Feret diameter and particle color detection in batch and single image processing pages.
 
 ### Changed
 * Updated image canvas in single image processing page to center smaller images.